renamed k8s packages
