+++ pkg/client/k8s/clientset.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1alpha1
+package k8s
 
 import (
 	"github.com/pkg/errors"
