+++ pkg/client/k8s/config.go
@@ -14,14 +14,15 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1alpha1
+package k8s
 
 import (
+	"strings"
+
 	"github.com/openebs/lib-csi/pkg/common/env"
 	"github.com/pkg/errors"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
-	"strings"
 )
 
 // ConfigGetter abstracts fetching of kubernetes client config
