+++ pkg/client/k8s/configmap_test.go
@@ -14,22 +14,32 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1alpha1
+package k8s
 
 import (
 	"testing"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func TestClientsetGet(t *testing.T) {
+// test if configmap implements ConfigMapGetter interface
+var _ ConfigMapGetter = &Configmap{}
+
+func TestConfigMapGet(t *testing.T) {
 	tests := map[string]struct {
-		iserr bool
+		namespace string
+		name      string
+		options   metav1.GetOptions
+		iserr     bool
 	}{
-		"101": {true},
+		"101": {"", "", metav1.GetOptions{}, true},
+		"102": {"default", "", metav1.GetOptions{}, true},
+		"103": {"default", "myconf", metav1.GetOptions{}, true},
 	}
 
 	for name, mock := range tests {
 		t.Run(name, func(t *testing.T) {
-			_, err := Clientset().Get()
+			_, err := ConfigMap(mock.namespace, mock.name).Get(mock.options)
 			if !mock.iserr && err != nil {
 				t.Fatalf("Test '%s' failed: expected no error: actual '%s'", name, err)
 			}
