+++ pkg/client/k8s/dynamic_test.go
@@ -5,7 +5,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-http://www.apache.org/licenses/LICENSE-2.0
+    http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,19 +14,28 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1alpha1
+package k8s
 
 import (
-	"github.com/pkg/errors"
-	"k8s.io/apimachinery/pkg/version"
+	"testing"
 )
 
-// GetServerVersion uses the client-go Discovery client to get the
-// kubernetes version struct
-func GetServerVersion() (*version.Info, error) {
-	cs, err := Clientset().Get()
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get apiserver version")
+// test if dynamic implements DynamicProvider interface
+var _ DynamicProvider = &DynamicStruct{}
+
+func TestDynamicProvider(t *testing.T) {
+	tests := map[string]struct {
+		iserr bool
+	}{
+		"101": {true},
 	}
-	return cs.Discovery().ServerVersion()
+
+	for name, mock := range tests {
+		t.Run(name, func(t *testing.T) {
+			_, err := Dynamic().Provide()
+			if !mock.iserr && err != nil {
+				t.Fatalf("Test '%s' failed: expected no error: actual '%s'", name, err)
+			}
+		})
+	}
 }
